requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
colorama==0.4.3
distlib==0.3.1
Django==3.1.2
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
filelock==3.0.12
flake8==3.8.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==2.0.1
Pillow==7.2.0
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
pylint==2.5.3
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
virtualenv==20.0.32
wrapt==1.12.1
